:sparkles: [MCH] o2-mch-dcs-ccdb cli can upload MCH/DCSConfig to CCDB
o2-mch-dcs-check-ccdb renamed to o2-mch-dcs-ccdb as it can now also put
things (for the moment MCH/DCSConfig datapoint configuration for
dcs-proxy) instead of just querying.
