+++ Detectors/MUON/MCH/Conditions/CMakeLists.txt
@@ -22,9 +22,9 @@ if(BUILD_TESTING)
     PUBLIC_LINK_LIBRARIES O2::MCHConditions)
 endif()
 
-o2_add_executable(dcs-check-ccdb
+o2_add_executable(dcs-ccdb
                   COMPONENT_NAME mch
-                  SOURCES src/dcs-check-ccdb.cxx
+                  SOURCES src/dcs-ccdb.cxx
                   PUBLIC_LINK_LIBRARIES O2::Framework O2::DetectorsDCS O2::MCHConditions O2::CCDB)
 
 o2_add_executable(mch-dcs-processor-workflow
